tests: community features

+++ src/node/status_node.py
@@ -27,17 +27,28 @@ class StatusNode:
         self.pubkey = pubkey
         self.process = None
         self.log_thread = None
+        self.capture_logs = True
         self.logs = []
 
-    def start(self):
+    def start(self, capture_logs=True):
+        self.capture_logs = capture_logs
         command = ["./status-cli", "serve", "-n", self.name, "-p", self.port]
         logger.info(f"Starting node with command: {command}")
         if self.pubkey:
             command += ["-a", self.pubkey]
 
         self.process = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
+        if self.capture_logs:
+            self._capture_logs()
+        self.api = StatusNodeRPC(self.port, self.name)
+        DS.nodes.append(self)
+
+    def serve_account(self, key_uuid):
+        command = ["./status-cli", "serve-account", "-n", self.name, "-kid", key_uuid]
+        logger.info(f"Serving node with command: {command}")
+        self.process = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, text=True)
         self._capture_logs()
-        self.api = StatusNodeRPC(self.port)
+        self.api = StatusNodeRPC(self.port, self.name)
         DS.nodes.append(self)
 
     def _capture_logs(self):
@@ -53,13 +64,14 @@ class StatusNode:
         self.log_thread.start()
         delay(2)  # Allow some time for the node to start and generate output
 
-    def stop(self):
+    def stop(self, remove_local_data=True):
         if self.process:
             logger.info(f"Stopping node with name: {self.name}")
             self.process.kill()
-            self.log_thread.join()  # Ensure log thread finishes
+            if self.capture_logs:
+                self.log_thread.join()  # Ensure log thread finishes
             node_dir = f"test-{self.name}"
-            if os.path.exists(node_dir):
+            if remove_local_data and os.path.exists(node_dir):
                 try:
                     shutil.rmtree(node_dir)
                 except Exception as ex:
@@ -133,6 +145,30 @@ class StatusNode:
     def send_group_chat_message(self, group_id, message):
         params = [{"id": group_id, "message": message}]
         return self.api.send_rpc_request("wakuext_sendGroupChatMessage", params)
+
+    def create_community(self, name):
+        params = [{"membership": 3, "name": name, "color": "#ffffff", "description": name}]
+        return self.api.send_rpc_request("wakuext_createCommunity", params)
+
+    def fetch_community(self, community_key):
+        params = [{"communityKey": community_key, "waitForResponse": True, "tryDatabase": True}]
+        return self.api.send_rpc_request("wakuext_fetchCommunity", params)
+
+    def request_to_join_community(self, community_id):
+        params = [{"communityId": community_id}]
+        return self.api.send_rpc_request("wakuext_requestToJoinCommunity", params)
+
+    def accept_request_to_join_community(self, request_to_join_id):
+        params = [{"id": request_to_join_id}]
+        return self.api.send_rpc_request("wakuext_acceptRequestToJoinCommunity", params)
+
+    def send_community_chat_message(self, chat_id, message):
+        params = [{"chatId": chat_id, "text": message, "contentType": 1}]
+        return self.api.send_rpc_request("wakuext_sendChatMessage", params)
+
+    def leave_community(self, community_id):
+        params = [community_id]
+        return self.api.send_rpc_request("wakuext_leaveCommunity", params)
 
     def random_node_name(self, length=10):
         allowed_chars = string.ascii_lowercase + string.digits + "_-"
